+++ src/rendering/fonts.h
@@ -8,4 +8,4 @@
 
 enum fontAlignment {LEFT_ALIGNMENT, CENTERED_ALIGNMET, RIGHT_ALIGNMENT};
 
-void drawText(appData_t *app, SDL_Texture *fontTexture, char *text, uint8_t x, uint8_t y, enum fontAlignment alignment);  
+void drawText(appData_t *app, SDL_Texture *fontTexture, char *text, uint_fast16_t x, uint_fast16_t y, enum fontAlignment alignment);
