+++ src/rendering/renderer.c
@@ -1,6 +1,6 @@
 #include "renderer.h"
 
-void blit(appData_t *app, SDL_Texture *texture, SDL_Rect *sourceRect, const int x, const int y, const int w, const int h) {
+void blit(appData_t *app, SDL_Texture *texture, SDL_Rect *sourceRect, const uint_fast16_t x, const uint_fast16_t y, const uint_fast16_t w, const uint_fast16_t h) {
   
   SDL_Rect destRect;
   destRect.x = x;
