+++ src/rendering/renderer.h
@@ -1,7 +1,8 @@
 #pragma once
 
+#include <stdint.h>
 #include "../sdlInclude.h"
 
 #include "../globalStructs.h"
 
-void blit(appData_t *app, SDL_Texture *texture, SDL_Rect *sourceRect, const int x, const int y, const int w, const int h); 
+void blit(appData_t *app, SDL_Texture *texture, SDL_Rect *sourceRect, const uint_fast16_t x, const uint_fast16_t y, const uint_fast16_t w, const uint_fast16_t h);
